+++ src/ypd/form/user_form.py
@@ -2,12 +2,6 @@ from flask_wtf import FlaskForm, Form
 from wtforms import BooleanField, FormField, PasswordField, StringField, SubmitField, IntegerField, RadioField, validators
 from wtforms.validators import Email, InputRequired, Length, EqualTo, DataRequired
 
-
-class TelephoneForm(Form):
-    country_code = IntegerField('Country Code', validators=[DataRequired()])
-    area_code    = IntegerField('Area Code/Exchange', validators=[DataRequired()])
-    number       = StringField('Number')
-
 class RegistrationForm(FlaskForm):
     username = StringField('Username', validators=[InputRequired(), Length(min=4, max=15)])
     password = PasswordField('Password', validators=[InputRequired(), Length(min=8, max=80)])
