+++ src/ypd/servlets/user_servlet.py
@@ -3,7 +3,7 @@ from flask_classy import FlaskView, route
 from flask_login import login_required, login_user, logout_user
 from sqlalchemy.exc import IntegrityError
 
-from ypd.model.flaskforms import LoginForm, RegistrationForm
+from ypd.form.user_form import LoginForm, RegistrationForm
 from ypd.model.user import User
 
 """A class that represents User creation routes"""
