+++ src/ypd/templates/project.html
@@ -18,15 +18,15 @@
         </a>
     </div>
 
-    <!-- % if project.can_be_modified_by(user) % -->
-    <div class="projectButton">
-        <a href="{{url_for('ProjectView:edit', id=project.id, is_provided=(project is provided))}}">
-            <button>
-                ✏️
-            </button>
-        </a>
-    </div>
-    <!-- % endif % -->
+    {% if project.can_be_modified_by(user) %}
+        <div class="projectButton">
+            <a href="{{url_for('ProjectView:edit', id=project.id, is_provided=(project is provided))}}">
+                <button>
+                    ✏️
+                </button>
+            </a>
+        </div>
+    {% endif %}
 
     <div class="bodyPadding">
         Company:
